refactor: split offline capture repository modules
